Added new stuff and restrictions

+++ app/__init__.py
@@ -105,16 +105,18 @@ def show_one_thing(id):
 def add_a_thing():
     # Get the data from the form
     name  = request.form.get("name")
-    price = request.form.get("price")
 
     # Sanitise the inputs
     name = html.escape(name)
-    price = html.escape(price)
+
+    #Get the user_id from the session 
+    user_id = session["user_id"]
+
 
     with connect_db() as client:
         # Add the thing to the DB
-        sql = "INSERT INTO things (name, price) VALUES (?, ?)"
-        values = [name, price]
+        sql = "INSERT INTO things (name, user_id) VALUES (?, ?)"
+        values = [name, user_id]
         client.execute(sql, values)
 
         # Go back to the home page
@@ -175,7 +177,7 @@ def login_user():
         #Check if we got a record
         if result.rows:
             #Yes, User exist
-            user = result.row[0]
+            user = result.rows[0]
             hash = user["password_hash"]
 
             #Check if passwords match
@@ -198,9 +200,13 @@ def login_user():
 #-----------------------------------------------------------
 @app.get("/delete/<int:id>")
 def delete_a_thing(id):
-    with connect_db() as client:
+    with connect_db() as client: 
+        # Get our user id from a session 
+        user_id = session["user_id"]
+
         # Delete the thing from the DB
-        sql = "DELETE FROM things WHERE id=?"
+        # Checking that we are the owner 
+        sql = "DELETE FROM things WHERE id=? AND user_id=?"
         values = [id]
         client.execute(sql, values)
 
@@ -208,4 +214,13 @@ def delete_a_thing(id):
         flash("Thing deleted", "warning")
         return redirect("/things")
 
-
+#-----------------------------------------------------------
+# Route for user logout
+#-----------------------------------------------------------
+    
+@app.get("/logout")
+def logout():
+    session.pop("user_id")
+    session.pop("user_name")
+    flash(" You have succesfully logged out yay!!!!")
+    return redirect("/")
